+++ lib/commands/fortnite/sac/sac.dart
@@ -1,6 +1,7 @@
 import "package:nyxx_commands/nyxx_commands.dart";
 import "view.dart";
 import "change.dart";
+import "earnings.dart";
 
 final ChatGroup affiliateCommand = ChatGroup(
   "sac",
@@ -8,6 +9,7 @@ final ChatGroup affiliateCommand = ChatGroup(
   children: [
     affiliateViewCommand,
     affiliateChangeCommand,
+    affiliateEarningsCommand,
   ],
   aliases: ["cc"],
   checks: [],